requirements.txt
# FastAPI e dependências principais (versões compatíveis com Python 3.8)
fastapi>=0.68.0,<0.100.0
uvicorn[standard]>=0.15.0,<0.25.0
pydantic>=1.8.0,<2.0.0

# Banco de dados (SQLite para desenvolvimento)
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<1.13.0

# Variáveis de ambiente
python-dotenv>=0.19.0

# Validação e serialização
email-validator>=1.1.0

# CORS
python-multipart>=0.0.5

# Desenvolvimento (opcional)
pytest>=6.0.0
pytest-asyncio>=0.15.0
httpx>=0.24.0